+++ SadConsole.GameHelpers.Windows/Properties/AssemblyInfo.cs
@@ -5,11 +5,11 @@ using System.Runtime.InteropServices;
 // General Information about an assembly is controlled through the following 
 // set of attributes. Change these attribute values to modify the information
 // associated with an assembly.
-[assembly: AssemblyTitle("SadConsole.Platforms.Windows.GameHelpers")]
+[assembly: AssemblyTitle("SadConsole.Platforms.Windows.Game")]
 [assembly: AssemblyDescription("")]
 [assembly: AssemblyConfiguration("")]
 [assembly: AssemblyCompany("")]
-[assembly: AssemblyProduct("SadConsole.Platforms.Windows.GameHelpers")]
+[assembly: AssemblyProduct("SadConsole.Platforms.Windows.Game")]
 [assembly: AssemblyCopyright("Copyright ©  2015")]
 [assembly: AssemblyTrademark("")]
 [assembly: AssemblyCulture("")]